Clarify Davidson coupling boundaries
