+++ src/mlx_atomistic/dft/_periodic_davidson_scheduler.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from collections import defaultdict
 from collections.abc import Callable, Sequence
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import Literal
 
 import mlx.core as mx
@@ -15,6 +15,7 @@ from mlx_atomistic.dft._compact import (
     _CompactBatchPolicy,
     _CompactLaneState,
 )
+from mlx_atomistic.dft._periodic_davidson_context import _FixedHamiltonianToken
 from mlx_atomistic.dft._periodic_davidson_planner import (
     _CompactBatchCapacity,
     _CompactSubmission,
@@ -27,84 +28,8 @@ from mlx_atomistic.dft._periodic_davidson_planner import (
 from mlx_atomistic.dft._periodic_execution import _detached_failure
 from mlx_atomistic.dft._periodic_hamiltonian import PeriodicKohnShamOperator
 from mlx_atomistic.dft._periodic_models import PeriodicDavidsonConfig
-from mlx_atomistic.dft._periodic_orthonormalization import (
-    _DAVIDSON_RANK_POLICY,
-    _Complex64RankPolicy,
-)
+from mlx_atomistic.dft._periodic_orthonormalization import _Complex64RankPolicy
 from mlx_atomistic.dft._runtime_observer import RuntimeObserver, add_observed_work
-from mlx_atomistic.dft.periodic_gth import PeriodicGTHNonlocalOperator
-
-
-def _hamiltonian_context(
-    operator: PeriodicKohnShamOperator,
-    config: PeriodicDavidsonConfig,
-    n_bands: int,
-    rank_policy: _Complex64RankPolicy,
-) -> tuple[object, ...]:
-    nonlocal_context = (
-        None
-        if operator.nonlocal_operator is None
-        else (
-            id(operator.nonlocal_operator),
-            operator.nonlocal_operator._context_identity,
-        )
-    )
-    potential = operator._effective_local_potential
-    return (
-        id(operator),
-        id(potential),
-        tuple(int(value) for value in potential.shape),
-        str(potential.dtype),
-        operator.basis.basis_fingerprint,
-        operator.basis.order_fingerprint,
-        operator.basis._layout.lane_id,
-        operator.basis.reciprocal_grid.fingerprint,
-        tuple(float(value) for value in operator.basis.kpoint_cartesian),
-        nonlocal_context,
-        "complex64-float32",
-        str(mx.default_device()),
-        config.max_iterations,
-        config.tolerance,
-        config.max_subspace_size,
-        config.preconditioner_floor,
-        n_bands,
-        "complex64-adaptive-choleskyqr-cgs2-mgs2-rank-v6",
-        rank_policy.relative_tolerance,
-    )
-
-
-@dataclass(frozen=True, eq=False)
-class _FixedHamiltonianToken:
-    """Solve-local identity that prevents paired H(V) from crossing contexts."""
-
-    context: tuple[object, ...]
-    nonce: object = field(default_factory=object, repr=False)
-
-    @classmethod
-    def create(
-        cls,
-        operator: PeriodicKohnShamOperator,
-        config: PeriodicDavidsonConfig,
-        n_bands: int,
-        rank_policy: _Complex64RankPolicy = _DAVIDSON_RANK_POLICY,
-    ) -> _FixedHamiltonianToken:
-        return cls(_hamiltonian_context(operator, config, n_bands, rank_policy))
-
-    def validate(
-        self,
-        operator: PeriodicKohnShamOperator,
-        config: PeriodicDavidsonConfig,
-        n_bands: int,
-        rank_policy: _Complex64RankPolicy = _DAVIDSON_RANK_POLICY,
-    ) -> None:
-        if self.context != _hamiltonian_context(
-            operator,
-            config,
-            n_bands,
-            rank_policy,
-        ):
-            msg = "Davidson H(V) token does not match the fixed Hamiltonian"
-            raise ValueError(msg)
 
 
 @dataclass(frozen=True)
@@ -213,18 +138,11 @@ class _DavidsonScheduler:
         ticket: _DavidsonApplicationTicket,
     ) -> tuple[object, ...]:
         layout = ticket.vectors.layout
-        nonlocal_operator = ticket.operator.nonlocal_operator
-        if nonlocal_operator is None:
-            nonlocal_context: object = None
-        elif isinstance(nonlocal_operator, PeriodicGTHNonlocalOperator):
-            nonlocal_context = ("gth", nonlocal_operator._context_identity)
-        else:
-            nonlocal_context = ("custom", id(nonlocal_operator))
         return (
             id(layout.reciprocal),
             layout.grid_shape,
             id(_DavidsonScheduler._observer(ticket)),
-            nonlocal_context,
+            ticket.operator._nonlocal_batch_compatibility_key(),
         )
 
     def _group_key(
